[5.2.x] draft: из зависимостей убран redux
