+++ CHANGELOG.md
@@ -59,12 +59,13 @@
 * 5.2.1 upd: конкретизированы типы TS
 * 5.2.1 upd: компоненты клиента переписаны на функциональные
 * 5.2.2 upd: переработанный глобальный стейт поставляется через контекст
-* 5.2.2 upd: дальнейший рефакторинг внешнего стейта, выделение презентационных компонентов
+* 5.2.2 upd: внешний стейт упрощен и разделен на recovery и common, компоненты разделены по файлам
+* 5.2.3 upd: убраны стейт-машины из комопнентов Catalog и Read
 
 # Запланированные изменения:
 * **БД**: перейти с MySQL на Postgres, сохранив демонстрационный дамп с ~1000 заметок
 * **Тесты**: юниты для JS, подготовить среду для E2E тестов, покрыть код автотестами
-* **React**: упростить контейнер со стейтом, задействовать Flux/Redux
+* **React**: упростить контейнер со стейтом, подумать над применимостью в клиенте Flux/Redux
 * **GitHub CI**: соединить пробные пайплайны в один, завязать публикацию образа на выпуск релиза 
                  и использовать для него артефакты билда сервиса и фронта
 * **GitHub CD**: подготовить варианты интеграции с высокодоступным кластером: сейчас сервис задеплоен на k3s
