+++ src/Rsse.Client/ClientApp/package.json
@@ -6,9 +6,7 @@
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-redux": "^9.1.0",
     "react-router-dom": "6.21.3",
-    "redux": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
   "devDependencies": {
