+++ src/Rsse.Client/ClientApp/src/jsx/components/create.checkbox.tsx
@@ -2,12 +2,12 @@
 import {Dispatch, SetStateAction, useContext} from "react";
 import {NoteResponseDto} from "../dto/request.response.dto";
 import {CommonContext} from "../common/context.provider";
-import {getStructuredTagsListResponse, getTagsCheckedUncheckedResponse} from "../common/dto.handlers";
+import {getStructuredTagsListResponse, getTagCheckedUncheckedResponse} from "../common/dto.handlers";
 import {ComponentMode} from "../common/state.handlers";
 
 export const CreateCheckbox = (props: {noteDto: NoteResponseDto, id: string, onClick: Dispatch<SetStateAction<NoteResponseDto|null>>}) => {
     const commonContext = useContext(CommonContext);
-    const checked = getTagsCheckedUncheckedResponse(props) === "checked";
+    const isChecked = getTagCheckedUncheckedResponse(props) === "checked";
 
     const getTagName = (i: number) => {
         return getStructuredTagsListResponse(props.noteDto)[i];
@@ -24,16 +24,17 @@ export const CreateCheckbox = (props: {noteDto: NoteResponseDto, id: string, onC
     };
 
     const loadNoteOnClick = (e: React.SyntheticEvent) => {
-        if (commonContext.componentMode == ComponentMode.ExtendedMode) {
-            let title = e.currentTarget.innerHTML.valueOf();
-            // item(1) это аттрибут about, в нём должен храниться id заметки, на который указывает данный чекбокс:
-            let id = e.currentTarget.attributes.item(1)?.nodeValue;
-            console.log(`Submitted & redirected: state: ${commonContext.componentMode} title: ${title} id: ${id}`);
-
+        // extended mode:
+        if (commonContext.componentMode == ComponentMode.Extended) {
+            const title = e.currentTarget.innerHTML.valueOf();
             const noteResponseDto = new NoteResponseDto();
+            // item(1) это аттрибут about, в нём должен храниться id заметки, на который указывает данный чекбокс:
+            const id = e.currentTarget.attributes.item(1)?.nodeValue;
+            console.log(`Submitted & redirected: state: ${commonContext.componentMode} title: ${title} id: ${id}`);
             // установка commonNoteID приведет к вызову редиректа после перерисовки CreateView:
             // commonNoteID также выставляется при сохранении новой заметки:
             noteResponseDto.commonNoteID = Number(id);
+            // const isClassicMode = isNaN(Number(id)); // косвенный признак состояния компонента
             props.onClick(noteResponseDto);
         }
     }
@@ -42,7 +43,7 @@ export const CreateCheckbox = (props: {noteDto: NoteResponseDto, id: string, onC
         <div id="checkboxStyle">
             <input name="chkButton" value={props.id} type="checkbox" id={props.id}
                    className="regular-checkbox"
-                   defaultChecked={checked}/>
+                   defaultChecked={isChecked}/>
             <label htmlFor={props.id}
                    onClick={loadNoteOnClick}
                    about={getTagId(Number(props.id))}>{getTagName(Number(props.id))}</label>
